[binding.velux] Bugfix: Handle empty list of items properly.     [binding.velux] Bugfix: HTTP/JSON communication.     [binding.velux] Bugfix: Aborting connection after sending error.     [binding.velux] Bugfix: Firmware revision properly returned.     [binding.velux] Bugfix: Close connections during binding shutdown/removal.     [binding.velux] Bugfix: String representation and handling of ip addresses.     [binding.velux] Bugfix: Invalid use of myJsonBridge eliminated in case of LAN/WLAN config.     [binding.velux] Bugfix: Initial refresh of data for item added, handling of BRIDGE_REFRESH_MSECS fixed.     [binding.velux] Bugfix: Refresh only readable items.     [binding.velux] Bugfix/enhancement: Adapted logging (more infos, less warnings).     [binding.velux] Enhancement: Comments for activation with OH2, more infos about debug settings in README, more detailed configuration in README added.     [binding.velux] Enhancement: More information during actuator state modification for debugging, more infos in loglevel trace.     [binding.velux] Enhancement: New channel timestamp.     [binding.velux] Enhancement: New channel Reload for retrieving all infos from bridge.     [binding.velux] Enhancement: Optional handling for inversed actuators.     [binding.velux] Enhancement: Eliminating usage of org.apache.commons.lang.StringUtils.isNotBlank.     [binding.velux] Enhancement: Proper handling of unknown config items.     [binding.velux] Enhancement: Provide infos about WLAN config.     [binding.velux] Enhancement: Better error handling for unknown scenes (virtual rollershutters).     [binding.velux] Enhancement: Intensive logging for recognition of Somfy devices.     [binding.velux] Enhancement: Handling bad serial numbers (Somfy).

+++ bundles/binding/org.openhab.binding.velux/src/main/java/org/openhab/binding/velux/bridge/slip/util/Packet.java
@@ -61,7 +61,7 @@ public class Packet {
     /**
      * Constructor: Create a {@link org.openhab.binding.velux.bridge.slip.util.Packet Packet} out of a sequence of
      * bytes.
-     * 
+     *
      * @param thisData Packet as Array of bytes.
      */
     public Packet(byte[] thisData) {
@@ -96,7 +96,7 @@ public class Packet {
     /**
      * Returns a part of the {@link org.openhab.binding.velux.bridge.slip.util.Packet Packet} as sequence of bytes
      * starting at position (n) up to the position (n+length-1).
-     * 
+     *
      * @param position Position (n) within the packet.
      * @param length Length of the intended slice as int.
      * @return <b>packet</b> of type Array-of-byte.
@@ -160,7 +160,7 @@ public class Packet {
      * <P>
      * Note: Big-endian LSB-0 encoding.
      * </P>
-     * 
+     *
      * @param position Position (n) within the packet.
      * @return <b>value</b> of type int.
      */
@@ -173,7 +173,7 @@ public class Packet {
      * <P>
      * Note: Big-endian LSB-0 encoding.
      * </P>
-     * 
+     *
      * @param position Position (n) within the packet.
      * @param value of type int.
      */
@@ -187,7 +187,7 @@ public class Packet {
      * <P>
      * Note: Big-endian LSB-0 encoding.
      * </P>
-     * 
+     *
      * @param position Position (n) within the packet.
      * @return <b>value</b> of type int.
      */
@@ -201,7 +201,7 @@ public class Packet {
      * <P>
      * Note: Big-endian LSB-0 encoding.
      * </P>
-     * 
+     *
      * @param position Position (n) within the packet.
      * @param value of type int.
      */
@@ -217,7 +217,7 @@ public class Packet {
      * <P>
      * Note: Any trailing null char will be eliminated.
      * </P>
-     * 
+     *
      * @param position Position (n) within the packet.
      * @param length Length of the intended slice as int.
      * @return <b>value</b> of type String.
@@ -232,7 +232,7 @@ public class Packet {
      * <P>
      * Note: The trailing null char will not be stored.
      * </P>
-     * 
+     *
      * @param position Position (n) within the packet.
      * @param text of type String.
      */
@@ -247,7 +247,7 @@ public class Packet {
 
     /**
      * Returns the hex char string representing the byte.
-     * 
+     *
      * @param oneByte of type byte to be converted.
      * @return <b>hexByteString</b> of type String.
      */
@@ -261,13 +261,13 @@ public class Packet {
 
     /**
      * Returns the dotted string representing an IP address.
-     * 
+     *
      * @param ipAddress of type int to be converted.
      * @return <b>ipAddressString</b> of type String.
      */
     public static String intToIPAddressString(int ipAddress) {
-        return String.format("%d.%d.%d.%d", (byte) ((ipAddress >>> 24) & 0xFF), (byte) ((ipAddress >>> 16) & 0xFF),
-                (byte) ((ipAddress >>> 8) & 0xFF), (byte) (ipAddress & 0xFF));
+        return String.format("%d.%d.%d.%d", ((ipAddress >>> 24) & 0xFF), ((ipAddress >>> 16) & 0xFF),
+                ((ipAddress >>> 8) & 0xFF), (ipAddress & 0xFF));
     }
 
 }
